+++ node/frontend/src/components/Canvas/Nodes/BuddySelect.tsx
@@ -9,6 +9,7 @@ import {
   DropdownMenuLabel,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
+import { availableLLMs } from "@/constants";
 import { useCanvas, useYDoc } from "@/hooks";
 import useBuddy from "@/hooks/useBuddy";
 import { BackendEntityType, YDocScope } from "@/types";
@@ -29,7 +30,7 @@ const BuddySelect = ({
   const isSkill = parent.type === BackendEntityType.SKILL;
   const buddyName = data?.buddy ? data?.buddy?.name : buddy?.name;
 
-  const setBuddy = (buddy: Buddy) => {
+  const setBuddy = (buddy?: Buddy) => {
     const node = nodesMap?.get(id);
     if (node) {
       nodesMap?.set(id, { ...node, data: { ...node?.data, buddy } });
@@ -58,10 +59,25 @@ const BuddySelect = ({
       {scope === YDocScope.READ_WRITE ? (
         <DropdownMenu modal={false}>
           <DropdownMenuTrigger asChild>{buddyBadge}</DropdownMenuTrigger>
-          <DropdownMenuContent align="end" className="text-neutral-700" sideOffset={8}>
+          <DropdownMenuContent
+            align="end"
+            className="max-h-[315px] overflow-auto text-neutral-700"
+            sideOffset={8}
+          >
             <DropdownMenuLabel className="border-b border-b-neutral-100 p-2 text-sm font-semibold">
               Buddy
             </DropdownMenuLabel>
+            <DropdownMenuItem
+              key="node-buddy-default"
+              className="flex cursor-pointer items-center text-sm font-medium capitalize
+                text-neutral-700"
+              onClick={() => setBuddy(undefined)}
+            >
+              <div className="mr-1 size-5">
+                {data.buddy === undefined && <Check className="size-4" />}
+              </div>
+              Default Skill Buddy ({buddy?.name})
+            </DropdownMenuItem>
             {buddies?.results?.map((b) => {
               return (
                 <DropdownMenuItem
@@ -72,11 +88,9 @@ const BuddySelect = ({
                   title={b?.system_message}
                 >
                   <div className="mr-1 size-5">
-                    {(b.id == data?.buddy?.id || (!data?.buddy && b.id == buddy?.id)) && (
-                      <Check className="size-4" />
-                    )}
+                    {b.id == data?.buddy?.id && <Check className="size-4" />}
                   </div>
-                  {b?.name} ({b?.model})
+                  {b?.name} ({b.model in availableLLMs ? availableLLMs[b.model] : "Unknown"})
                 </DropdownMenuItem>
               );
             })}