+++ node/frontend/src/components/SkillsRunner/Skill.tsx
@@ -1,0 +1,92 @@
+import { Skill as SkillType } from "@coordnet/core";
+import clsx from "clsx";
+import { Bot, Check, ChevronDown } from "lucide-react";
+import { Link, useParams } from "react-router-dom";
+
+import line from "@/assets/line-1.svg";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu";
+import { availableLLMs } from "@/constants";
+import useBuddy from "@/hooks/useBuddy";
+
+import SkillCard from "../Skills/SkillCard";
+
+export const Skill = ({ skill }: { skill?: SkillType }) => {
+  const { buddy, buddies, buddyId, setBuddyId } = useBuddy();
+  const { skillId, versionId, runId } = useParams();
+
+  const skillCard = skill ? (
+    <SkillCard
+      skill={skill}
+      disableInteraction={true}
+      className="z-1 relative !h-52 w-48 shadow-lg"
+    />
+  ) : (
+    <></>
+  );
+
+  return (
+    <div className="mx-auto">
+      <h2 className="mb-4 text-xl font-semibold text-neutral-400">Skill Card</h2>
+      <div className="flex flex-row gap-6">
+        <div className="relative">
+          {runId ? (
+            <Link to={`/skills/${skillId}/versions/${versionId}/runs/${runId}`} target="_blank">
+              {skillCard}
+            </Link>
+          ) : (
+            skillCard
+          )}
+        </div>
+        <img src={line} alt="line" className="hidden flex-shrink-0 md:block" />
+      </div>
+      <DropdownMenu modal={false}>
+        <DropdownMenuTrigger asChild disabled={Boolean(runId)}>
+          <div
+            className={clsx(
+              `mt-4 flex w-48 items-center gap-2 rounded-full border border-neutral-200 bg-white
+              px-3 py-2.5`,
+              !runId && "cursor-pointer"
+            )}
+          >
+            <Bot className="size-4 text-neutral-500" />
+            <div
+              className="flex-1 justify-start truncate text-sm font-medium text-neutral-900"
+              title={`${buddy?.name} (${buddy?.model})`}
+            >
+              {buddy?.name} ({buddy?.model})
+            </div>
+            {!runId && <ChevronDown className="size-4 text-neutral-500" />}
+          </div>
+        </DropdownMenuTrigger>
+        <DropdownMenuContent
+          align="end"
+          className="max-h-[315px] overflow-auto text-neutral-700"
+          sideOffset={8}
+        >
+          <DropdownMenuLabel className="border-b border-b-neutral-100 p-2 text-sm font-semibold">
+            Buddy
+          </DropdownMenuLabel>
+          {buddies?.results?.map((b) => {
+            return (
+              <DropdownMenuItem
+                key={`node-buddy-${b.id}`}
+                className="flex cursor-pointer items-center text-sm font-medium text-neutral-700"
+                onClick={() => setBuddyId(b.id)}
+                title={b?.system_message}
+              >
+                <div className="mr-1 size-5">{b.id == buddyId && <Check className="size-4" />}</div>
+                {b?.name} ({b.model in availableLLMs ? availableLLMs[b.model] : "Unknown"})
+              </DropdownMenuItem>
+            );
+          })}
+        </DropdownMenuContent>
+      </DropdownMenu>
+    </div>
+  );
+};