+++ node/frontend/src/hooks/useRunSkill.ts
@@ -4,6 +4,7 @@ import {
   ExecutionContext,
   ExecutionPlan,
   NodeType,
+  processTasks,
   RunStatus,
   skillYdocToJson,
 } from "@coordnet/core";
@@ -72,7 +73,6 @@ export const useRunSkill = () => {
 
     // Go to the run page and execute it
     await executeSkillRun(run.id);
-    // window.location.href = `/skills/${parent.id}${versionId ? `/versions/${versionId}` : ""}/runs/${run.id}`;
     navigate(`/skills/${parent.id}${versionId ? `/versions/${versionId}` : ""}/runs/${run.id}`);
   }, [buddy, inputNodes.length, navigate, nodes, parent.id, scope, skillId, spaceYDoc, versionId]);
 
@@ -86,6 +86,10 @@ export const useRunSkill = () => {
       toast.error("Output node not found");
       return;
     }
+    if (!skillId) {
+      toast.error("Skill ID not found");
+      return;
+    }
 
     const canvas = createCanvas(nodes, edges);
     const context: ExecutionContext = {
@@ -96,8 +100,8 @@ export const useRunSkill = () => {
     };
     createTasks(canvas, context);
 
-    // return processTasks(context, buddy, spaceYDoc, skill, skillNodesMap, nodesMap, cancelRef, true);
-  }, [buddy, edges, nodes]);
+    return await processTasks(context, buddy, skillId, spaceYDoc, true);
+  }, [buddy, edges, nodes, skillId, spaceYDoc]);
 
   return { runSkill, prepareExecutionPlan, status, error };
 };